Need to learn more about references, pointers, adresses, smart pointers etc.

+++ list.h
@@ -9,7 +9,7 @@ public:
 	int id; // list id
 	std::string name; // name of list
 
-	std::vector<int> vector{ std::vector<int>() }; // just trying with 10
+	std::vector<unsigned long long> vector{ std::vector<unsigned long long>() }; // just trying with 10
 	//	2. I guess I need to store the ptr to the task in the array??
